Initial UI
This is the first steps for the website. There is a master navigation bar on the top of each page
homePage contains a title for the website and a search bar to search events
create has a form attached with a submit button to create an event
viewEvents is where people can view the events they have created or have entered the lottery of. has an email input box